This change fixes a possible stack overflow when running the test.
Normally, window.requestAnimationFrame(callback) sets up an
asynchronous timer that will trigger the given callback after some
delay. The previous version of this test replaced
window.requestAnimationFrame with a function that calls the
callback immediately. When the code under test requests a new
animation frame unconditionally, this results in immediate
unbounded recursion and a stack overflow. (Unconditional request
of a new animation frame is useful when using
window.cancelAnimationFrame to control termination of the
animation.)

Replacing window.requestAnimationFrame does not appear to be
necessary to test whether moveRock checks for collisions. The
simplest solution seems to be to remove the reassignment of
window.requestAnimationFrame.