This fixes an infinite loop when running tests for moveRock().
The main issue here was using setInterval instead of setTimeout.
window.requestAnimationFrame is expected to call the callback a
single time, so if we are replacing its implementation we should use
setTimeout which calls the callback once, instead of setInterval
which calls the callback indefinitely.

We also need to redefine window.cancelAnimationFrame to provide a
compatible cancellation function for the redefined
requestAnimationFrame. We will also return the setTimeout id from
the new requestAnimationFrame so the caller can cancel the correct
timeout.

We must wait for the rock to fall to the bottom of the game screen
before we know if the test will pass. Apparently the previous version
of the test took < 50ms for this to happen. This fixed version of the
test takes longer. Empirically, it takes about 275ms for the rock to
fall on my machine, so I've used 300ms to make the test pass.